+++ bigml/tests/create_prediction_steps.py
@@ -210,6 +210,18 @@ def i_create_a_logistic_prediction(step, data=None):
     world.prediction = resource['object']
     world.predictions.append(resource['resource'])
 
+def i_create_a_deepnet_prediction(step, data=None):
+    if data is None:
+        data = "{}"
+    deepnet = world.deepnet['resource']
+    data = json.loads(data)
+    resource = world.api.create_prediction(deepnet, data)
+    world.status = resource['code']
+    eq_(world.status, HTTP_CREATED)
+    world.location = resource['location']
+    world.prediction = resource['object']
+    world.predictions.append(resource['resource'])
+
 def the_logistic_probability_is(step, probability):
     for [prediction, remote_probability] in world.prediction['probabilities']:
         if prediction == world.prediction['output']: