+++ bigml/tests/test_33_compare_predictions.py
@@ -339,3 +339,49 @@ class TestComparePrediction(object):
             prediction_create.create_local_ensemble_proportional_prediction_with_confidence(self, example[4])
             prediction_compare.the_local_ensemble_prediction_is(self, example[6])
             prediction_compare.the_local_prediction_confidence_is(self, example[7])
+
+    def test_scenario8(self):
+        """
+            Scenario: Successfully comparing predictions for deepnets:
+                Given I create a data source uploading a "<data>" file
+                And I wait until the source is ready less than <time_1> secs
+                And I create a dataset
+                And I wait until the dataset is ready less than <time_2> secs
+                And I create a deepnet with objective "<objective>" and "<params>"
+                And I wait until the deepnet is ready less than <time_3> secs
+                And I create a local deepnet
+                When I create a prediction for "<data_input>"
+                Then the prediction for "<objective>" is "<prediction>"
+                And I create a local prediction for "<data_input>"
+                Then the local prediction is "<prediction>"
+
+                Examples:
+                | data             | time_1  | time_2 | time_3 | data_input                             | objective | prediction  | params,
+            ['data/iris.csv', '10', '10', '120', '{"petal length": 6, "petal width": 2}', '000004', 'Iris-virginica', '{"number_of_models": 5}'],
+            ['data/iris.csv', '10', '10', '120', '{"petal length": 4, "petal width": 1.5}', '000004', 'Iris-versicolor', '{"number_of_models": 5}'],
+            ['data/grades.csv', '10', '10', '120', '{"Midterm": 20}', '000005', 46.261364, '{"number_of_models": 5}'],
+            ['data/iris.csv', '10', '10', '120', '{"petal width": 0.5}', '000004', 'Iris-setosa', '{"boosting": {"iterations": 5}, "number_of_models": 5}'],
+            ['data/iris.csv', '10', '10', '120', '{"petal length": 6, "petal width": 2}', '000004', 'Iris-virginica', '{"boosting": {"iterations": 5}, "number_of_models": 5}'],
+            ['data/iris.csv', '10', '10', '120', '{"petal length": 4, "petal width": 1.5}', '000004', 'Iris-versicolor', '{"boosting": {"iterations": 5}, "number_of_models": 5}'],
+
+            ['data/iris.csv', '10', '10', '120', '{"petal width": 0.5}', '000004', 'Iris-versicolor', '{}']
+
+
+        """
+        examples = [
+            ['data/iris.csv', '10', '50', '30000', '{"petal width": 4}', '000004', 'Iris-virginica', '{}']]
+        show_doc(self.test_scenario8, examples)
+
+        for example in examples:
+            print "\nTesting with:\n", example
+            source_create.i_upload_a_file(self, example[0])
+            source_create.the_source_is_finished(self, example[1])
+            dataset_create.i_create_a_dataset(self)
+            dataset_create.the_dataset_is_finished_in_less_than(self, example[2])
+            model_create.i_create_a_deepnet_with_objective_and_params(self, example[5], example[7])
+            model_create.the_deepnet_is_finished_in_less_than(self, example[3])
+            prediction_compare.i_create_a_local_deepnet(self)
+            prediction_create.i_create_a_deepnet_prediction(self, example[4])
+            prediction_create.the_prediction_is(self, example[5], example[6])
+            prediction_compare.i_create_a_local_deepnet_prediction(self, example[4])
+            prediction_compare.the_local_prediction_is(self, example[6])